cleanup: Move "api" folder outside of "components"
Since it's not really a component.
