+++ src/components/panels/EnterpriseListPanelComponent.js
@@ -3,7 +3,7 @@
 import React from 'react';
 import { browserHistory } from 'react-router';
 
-import api from '../api/api.js';
+import api from '../../api/api.js';
 
 require('styles/panels/EnterpriseListPanel.scss');
